Separate Model reference methods in a trait (#679)
* [update] separate reference methods in a trait
